chore: dag blocks processing refactor
