+++ libraries/cli/include/cli/config_jsons/testnet.json
@@ -51,7 +51,6 @@
   },
   "max_transactions_pool_warn": 0,
   "max_transactions_pool_drop": 0,
-  "max_block_queue_warn": 0,
   "db_config": {
     "db_snapshot_each_n_pbft_block": 10000,
     "db_max_snapshots": 5
