+++ libraries/core_libs/consensus/include/dag/block_proposer.hpp
@@ -7,6 +7,7 @@
 
 #include "boost/thread.hpp"
 #include "config/config.hpp"
+#include "dag/dag_manager.hpp"
 #include "logger/logger.hpp"
 #include "network/network.hpp"
 
@@ -16,7 +17,6 @@ namespace taraxa {
  * @{
  */
 
-class DagManager;
 class TransactionManager;
 class FullNode;
 class BlockProposer;
@@ -52,13 +52,10 @@ class ProposeModelFace {
 class SortitionPropose : public ProposeModelFace {
  public:
   SortitionPropose(addr_t node_addr, std::shared_ptr<DbStorage> db, std::shared_ptr<DagManager> dag_mgr,
-                   std::shared_ptr<DagBlockManager> dag_blk_mgr, std::shared_ptr<TransactionManager> trx_mgr)
-      : db_(std::move(db)),
-        dag_mgr_(std::move(dag_mgr)),
-        dag_blk_mgr_(std::move(dag_blk_mgr)),
-        trx_mgr_(std::move(trx_mgr)) {
+                   std::shared_ptr<TransactionManager> trx_mgr)
+      : db_(std::move(db)), dag_mgr_(std::move(dag_mgr)), trx_mgr_(std::move(trx_mgr)) {
     LOG_OBJECTS_CREATE("PR_MDL");
-    LOG(log_nf_) << "Set sortition DAG block proposal" << dag_blk_mgr_->sortitionParamsManager().getSortitionParams();
+    LOG(log_nf_) << "Set sortition DAG block proposal" << dag_mgr_->sortitionParamsManager().getSortitionParams();
     // Add a random component in proposing stale blocks so that not all nodes propose stale blocks at the same time
     // This will make stale block be proposed after waiting random interval between 2 and 20 seconds
     max_num_tries_ += (node_addr_[0] % (10 * max_num_tries_));
@@ -75,7 +72,6 @@ class SortitionPropose : public ProposeModelFace {
   util::ThreadPool executor_{1};
   std::shared_ptr<DbStorage> db_;
   std::shared_ptr<DagManager> dag_mgr_;
-  std::shared_ptr<DagBlockManager> dag_blk_mgr_;
   std::shared_ptr<TransactionManager> trx_mgr_;
 
   LOG_OBJECTS_DEFINE
@@ -93,20 +89,18 @@ class SortitionPropose : public ProposeModelFace {
 class BlockProposer : public std::enable_shared_from_this<BlockProposer> {
  public:
   BlockProposer(BlockProposerConfig const& bp_config, std::shared_ptr<DagManager> dag_mgr,
-                std::shared_ptr<TransactionManager> trx_mgr, std::shared_ptr<DagBlockManager> dag_blk_mgr,
-                std::shared_ptr<FinalChain> final_chain, std::shared_ptr<DbStorage> db, addr_t node_addr,
-                secret_t node_sk, vrf_sk_t vrf_sk)
+                std::shared_ptr<TransactionManager> trx_mgr, std::shared_ptr<FinalChain> final_chain,
+                std::shared_ptr<DbStorage> db, addr_t node_addr, secret_t node_sk, vrf_sk_t vrf_sk)
       : bp_config_(bp_config),
         dag_mgr_(std::move(dag_mgr)),
         trx_mgr_(std::move(trx_mgr)),
-        dag_blk_mgr_(std::move(dag_blk_mgr)),
         final_chain_(std::move(final_chain)),
         db_(std::move(db)),
         node_addr_(node_addr),
         node_sk_(node_sk),
         vrf_sk_(vrf_sk) {
     LOG_OBJECTS_CREATE("PR_MDL");
-    propose_model_ = std::make_unique<SortitionPropose>(node_addr, db_, dag_mgr_, dag_blk_mgr_, trx_mgr_);
+    propose_model_ = std::make_unique<SortitionPropose>(node_addr, db_, dag_mgr_, trx_mgr_);
     total_trx_shards_ = std::max((unsigned int)bp_config_.shard, 1u);
     auto addr = std::stoull(node_addr.toString().substr(0, 6).c_str(), NULL, 16);
     my_trx_shard_ = addr % bp_config_.shard;
@@ -179,7 +173,6 @@ class BlockProposer : public std::enable_shared_from_this<BlockProposer> {
   uint16_t my_trx_shard_;
   std::shared_ptr<DagManager> dag_mgr_;
   std::shared_ptr<TransactionManager> trx_mgr_;
-  std::shared_ptr<DagBlockManager> dag_blk_mgr_;
   std::shared_ptr<FinalChain> final_chain_;
   std::shared_ptr<DbStorage> db_;
   std::shared_ptr<std::thread> proposer_worker_;
