+++ libraries/core_libs/consensus/include/pbft/pbft_manager.hpp
@@ -6,6 +6,7 @@
 #include "common/types.hpp"
 #include "common/vrf_wrapper.hpp"
 #include "config/config.hpp"
+#include "final_chain/final_chain.hpp"
 #include "key_manager/key_manager.hpp"
 #include "logger/logger.hpp"
 #include "network/network.hpp"
@@ -72,10 +73,9 @@ class PbftManager : public std::enable_shared_from_this<PbftManager> {
   PbftManager(const PbftConfig &conf, const blk_hash_t &dag_genesis_block_hash, addr_t node_addr,
               std::shared_ptr<DbStorage> db, std::shared_ptr<PbftChain> pbft_chain,
               std::shared_ptr<VoteManager> vote_mgr, std::shared_ptr<NextVotesManager> next_votes_mgr,
-              std::shared_ptr<DagManager> dag_mgr, std::shared_ptr<DagBlockManager> dag_blk_mgr,
-              std::shared_ptr<TransactionManager> trx_mgr, std::shared_ptr<FinalChain> final_chain,
-              std::shared_ptr<KeyManager> key_manager, secret_t node_sk, vrf_sk_t vrf_sk,
-              uint32_t max_levels_per_period = kMaxLevelsPerPeriod);
+              std::shared_ptr<DagManager> dag_mgr, std::shared_ptr<TransactionManager> trx_mgr,
+              std::shared_ptr<FinalChain> final_chain, std::shared_ptr<KeyManager> key_manager, secret_t node_sk,
+              vrf_sk_t vrf_sk, uint32_t max_levels_per_period = kMaxLevelsPerPeriod);
   ~PbftManager();
   PbftManager(const PbftManager &) = delete;
   PbftManager(PbftManager &&) = delete;
@@ -584,7 +584,6 @@ class PbftManager : public std::enable_shared_from_this<PbftManager> {
   std::shared_ptr<VoteManager> vote_mgr_;
   std::shared_ptr<DagManager> dag_mgr_;
   std::weak_ptr<Network> network_;
-  std::shared_ptr<DagBlockManager> dag_blk_mgr_;
   std::shared_ptr<TransactionManager> trx_mgr_;
   std::shared_ptr<FinalChain> final_chain_;
   std::shared_ptr<KeyManager> key_manager_;
