+++ libraries/core_libs/consensus/src/pbft/pbft_manager.cpp
@@ -29,16 +29,14 @@ using vrf_output_t = vrf_wrapper::vrf_output_t;
 PbftManager::PbftManager(const PbftConfig &conf, const blk_hash_t &dag_genesis_block_hash, addr_t node_addr,
                          std::shared_ptr<DbStorage> db, std::shared_ptr<PbftChain> pbft_chain,
                          std::shared_ptr<VoteManager> vote_mgr, std::shared_ptr<NextVotesManager> next_votes_mgr,
-                         std::shared_ptr<DagManager> dag_mgr, std::shared_ptr<DagBlockManager> dag_blk_mgr,
-                         std::shared_ptr<TransactionManager> trx_mgr, std::shared_ptr<FinalChain> final_chain,
-                         std::shared_ptr<KeyManager> key_manager, secret_t node_sk, vrf_sk_t vrf_sk,
-                         uint32_t max_levels_per_period)
+                         std::shared_ptr<DagManager> dag_mgr, std::shared_ptr<TransactionManager> trx_mgr,
+                         std::shared_ptr<FinalChain> final_chain, std::shared_ptr<KeyManager> key_manager,
+                         secret_t node_sk, vrf_sk_t vrf_sk, uint32_t max_levels_per_period)
     : db_(std::move(db)),
       next_votes_manager_(std::move(next_votes_mgr)),
       pbft_chain_(std::move(pbft_chain)),
       vote_mgr_(std::move(vote_mgr)),
       dag_mgr_(std::move(dag_mgr)),
-      dag_blk_mgr_(std::move(dag_blk_mgr)),
       trx_mgr_(std::move(trx_mgr)),
       final_chain_(std::move(final_chain)),
       key_manager_(std::move(key_manager)),
@@ -1280,7 +1278,7 @@ blk_hash_t PbftManager::proposePbftBlock_() {
   u256 total_weight = 0;
   uint32_t dag_blocks_included = 0;
   for (auto const &blk_hash : dag_block_order) {
-    auto dag_blk = dag_blk_mgr_->getDagBlock(blk_hash);
+    auto dag_blk = dag_mgr_->getDagBlock(blk_hash);
     if (!dag_blk) {
       LOG(log_er_) << "DAG anchor block hash " << dag_block_hash << " getDagBlock failed in propose for block "
                    << blk_hash;
@@ -1321,7 +1319,7 @@ blk_hash_t PbftManager::proposePbftBlock_() {
     std::unordered_set<trx_hash_t> trx_set;
     std::vector<trx_hash_t> transactions_to_query;
     for (auto const &dag_blk_hash : dag_block_order) {
-      auto dag_block = dag_blk_mgr_->getDagBlock(dag_blk_hash);
+      auto dag_block = dag_mgr_->getDagBlock(dag_blk_hash);
       assert(dag_block);
       for (auto const &trx_hash : dag_block->getTrxs()) {
         if (trx_set.insert(trx_hash).second) {
@@ -1513,7 +1511,7 @@ std::pair<vec_blk_t, bool> PbftManager::compareBlocksAndRewardVotes_(std::shared
   period_data_.clear();
   period_data_.dag_blocks.reserve(dag_blocks_order.size());
   for (auto const &dag_blk_hash : dag_blocks_order) {
-    auto dag_block = dag_blk_mgr_->getDagBlock(dag_blk_hash);
+    auto dag_block = dag_mgr_->getDagBlock(dag_blk_hash);
     assert(dag_block);
     for (auto const &trx_hash : dag_block->getTrxs()) {
       if (trx_set.insert(trx_hash).second) {
@@ -1531,7 +1529,12 @@ std::pair<vec_blk_t, bool> PbftManager::compareBlocksAndRewardVotes_(std::shared
   }
 
   const auto transactions = trx_mgr_->getNonfinalizedTrx(non_finalized_transactions, true /*sorted*/);
+  if (transactions.size() < non_finalized_transactions.size()) {
+    return std::make_pair(std::move(dag_blocks_order), false);
+  }
+
   non_finalized_transactions.clear();
+
   period_data_.transactions.reserve(transactions.size());
   for (const auto &trx : transactions) {
     non_finalized_transactions.push_back(trx->getHash());
@@ -1652,7 +1655,7 @@ void PbftManager::finalize_(PeriodData &&period_data, std::vector<h256> &&finali
           return;
         }
 
-        auto anchor = dag_blk_mgr_->getDagBlock(anchor_hash);
+        auto anchor = dag_mgr_->getDagBlock(anchor_hash);
         if (!anchor) {
           LOG(log_er_) << "DB corrupted - Cannot find anchor block: " << anchor_hash << " in DB.";
           assert(false);
@@ -1704,8 +1707,8 @@ bool PbftManager::pushPbftBlock_(PeriodData &&period_data, std::vector<std::shar
 
   // pass pbft with dag blocks and transactions to adjust difficulty
   if (period_data.pbft_blk->getPivotDagBlockHash() != NULL_BLOCK_HASH) {
-    dag_blk_mgr_->sortitionParamsManager().pbftBlockPushed(period_data, batch,
-                                                           pbft_chain_->getPbftChainSizeExcludingEmptyPbftBlocks() + 1);
+    dag_mgr_->sortitionParamsManager().pbftBlockPushed(period_data, batch,
+                                                       pbft_chain_->getPbftChainSizeExcludingEmptyPbftBlocks() + 1);
   }
   {
     // This makes sure that no DAG block or transaction can be added or change state in transaction and dag manager when
