+++ libraries/core_libs/network/include/network/tarcap/packets_handlers/dag_block_packet_handler.hpp
@@ -15,8 +15,8 @@ class DagBlockPacketHandler final : public ExtSyncingPacketHandler {
   DagBlockPacketHandler(std::shared_ptr<PeersState> peers_state, std::shared_ptr<PacketsStats> packets_stats,
                         std::shared_ptr<PbftSyncingState> pbft_syncing_state, std::shared_ptr<PbftChain> pbft_chain,
                         std::shared_ptr<PbftManager> pbft_mgr, std::shared_ptr<DagManager> dag_mgr,
-                        std::shared_ptr<DagBlockManager> dag_blk_mgr, std::shared_ptr<TransactionManager> trx_mgr,
-                        std::shared_ptr<DbStorage> db, std::shared_ptr<TestState> test_state, const addr_t &node_addr);
+                        std::shared_ptr<TransactionManager> trx_mgr, std::shared_ptr<DbStorage> db,
+                        std::shared_ptr<TestState> test_state, const addr_t &node_addr);
 
   void sendBlock(dev::p2p::NodeID const &peer_id, DagBlock block, const SharedTransactions &trxs);
   void onNewBlockReceived(DagBlock &&block, const std::shared_ptr<TaraxaPeer> &peer = nullptr);
