+++ libraries/core_libs/network/include/network/tarcap/packets_handlers/dag_sync_packet_handler.hpp
@@ -13,8 +13,8 @@ class DagSyncPacketHandler final : public ExtSyncingPacketHandler {
   DagSyncPacketHandler(std::shared_ptr<PeersState> peers_state, std::shared_ptr<PacketsStats> packets_stats,
                        std::shared_ptr<PbftSyncingState> pbft_syncing_state, std::shared_ptr<PbftChain> pbft_chain,
                        std::shared_ptr<PbftManager> pbft_mgr, std::shared_ptr<DagManager> dag_mgr,
-                       std::shared_ptr<TransactionManager> trx_mgr, std::shared_ptr<DagBlockManager> dag_blk_mgr,
-                       std::shared_ptr<DbStorage> db, const addr_t& node_addr);
+                       std::shared_ptr<TransactionManager> trx_mgr, std::shared_ptr<DbStorage> db,
+                       const addr_t& node_addr);
 
   // Packet type that is processed by this handler
   static constexpr SubprotocolPacketType kPacketType_ = SubprotocolPacketType::DagSyncPacket;
