+++ libraries/core_libs/network/include/network/tarcap/packets_handlers/get_dag_sync_packet_handler.hpp
@@ -4,7 +4,6 @@
 
 namespace taraxa {
 class DagManager;
-class DagBlockManager;
 class DbStorage;
 class TransactionManager;
 }  // namespace taraxa
@@ -15,8 +14,7 @@ class GetDagSyncPacketHandler final : public PacketHandler {
  public:
   GetDagSyncPacketHandler(std::shared_ptr<PeersState> peers_state, std::shared_ptr<PacketsStats> packets_stats,
                           std::shared_ptr<TransactionManager> trx_mgr, std::shared_ptr<DagManager> dag_mgr,
-                          std::shared_ptr<DagBlockManager> dag_blk_mgr, std::shared_ptr<DbStorage> db,
-                          const addr_t& node_addr);
+                          std::shared_ptr<DbStorage> db, const addr_t& node_addr);
 
   void sendBlocks(const dev::p2p::NodeID& peer_id, std::vector<std::shared_ptr<DagBlock>>&& blocks,
                   SharedTransactions&& transactions, uint64_t request_period, uint64_t period);
@@ -30,7 +28,6 @@ class GetDagSyncPacketHandler final : public PacketHandler {
 
   std::shared_ptr<TransactionManager> trx_mgr_;
   std::shared_ptr<DagManager> dag_mgr_;
-  std::shared_ptr<DagBlockManager> dag_blk_mgr_;
   std::shared_ptr<DbStorage> db_;
 };
