+++ libraries/core_libs/network/include/network/tarcap/packets_handlers/pbft_block_packet_handler.hpp
@@ -6,7 +6,6 @@ namespace taraxa {
 class PbftBlock;
 class PbftChain;
 class PbftManager;
-class DagBlockManager;
 class VoteManager;
 }  // namespace taraxa
