+++ libraries/core_libs/network/include/network/tarcap/packets_handlers/pbft_sync_packet_handler.hpp
@@ -10,9 +10,8 @@ class PbftSyncPacketHandler final : public ExtSyncingPacketHandler {
   PbftSyncPacketHandler(std::shared_ptr<PeersState> peers_state, std::shared_ptr<PacketsStats> packets_stats,
                         std::shared_ptr<PbftSyncingState> pbft_syncing_state, std::shared_ptr<PbftChain> pbft_chain,
                         std::shared_ptr<PbftManager> pbft_mgr, std::shared_ptr<DagManager> dag_mgr,
-                        std::shared_ptr<DagBlockManager> dag_blk_mgr, std::shared_ptr<VoteManager> vote_mgr,
-                        std::shared_ptr<util::ThreadPool> periodic_events_tp, std::shared_ptr<DbStorage> db,
-                        size_t network_sync_level_size, const addr_t& node_addr);
+                        std::shared_ptr<VoteManager> vote_mgr, std::shared_ptr<util::ThreadPool> periodic_events_tp,
+                        std::shared_ptr<DbStorage> db, size_t network_sync_level_size, const addr_t& node_addr);
 
   void handleMaliciousSyncPeer(dev::p2p::NodeID const& id);
