+++ libraries/core_libs/network/include/network/tarcap/packets_handlers/status_packet_handler.hpp
@@ -13,9 +13,8 @@ class StatusPacketHandler final : public ExtSyncingPacketHandler {
   StatusPacketHandler(std::shared_ptr<PeersState> peers_state, std::shared_ptr<PacketsStats> packets_stats,
                       std::shared_ptr<PbftSyncingState> pbft_syncing_state, std::shared_ptr<PbftChain> pbft_chain,
                       std::shared_ptr<PbftManager> pbft_mgr, std::shared_ptr<DagManager> dag_mgr,
-                      std::shared_ptr<DagBlockManager> dag_blk_mgr, std::shared_ptr<NextVotesManager> next_votes_mgr,
-                      std::shared_ptr<DbStorage> db, uint64_t conf_network_id, h256 genesis_hash,
-                      const addr_t& node_addr);
+                      std::shared_ptr<NextVotesManager> next_votes_mgr, std::shared_ptr<DbStorage> db,
+                      uint64_t conf_network_id, h256 genesis_hash, const addr_t& node_addr);
 
   bool sendStatus(const dev::p2p::NodeID& node_id, bool initial);
   void sendStatusToPeers();
