+++ libraries/core_libs/network/include/network/tarcap/packets_handlers/transaction_packet_handler.hpp
@@ -5,7 +5,6 @@
 #include "transaction/transaction.hpp"
 
 namespace taraxa {
-class DagBlockManager;
 class TransactionManager;
 enum class TransactionStatus;
 }  // namespace taraxa
@@ -17,8 +16,8 @@ class TestState;
 class TransactionPacketHandler final : public PacketHandler {
  public:
   TransactionPacketHandler(std::shared_ptr<PeersState> peers_state, std::shared_ptr<PacketsStats> packets_stats,
-                           std::shared_ptr<TransactionManager> trx_mgr, std::shared_ptr<DagBlockManager> dag_blk_mgr,
-                           std::shared_ptr<TestState> test_state, const addr_t& node_addr);
+                           std::shared_ptr<TransactionManager> trx_mgr, std::shared_ptr<TestState> test_state,
+                           const addr_t& node_addr);
 
   void onNewTransactions(std::vector<std::pair<std::shared_ptr<Transaction>, TransactionStatus>>&& transactions);
   void sendTransactions(dev::p2p::NodeID const& peer_id, std::vector<taraxa::bytes> const& transactions);
@@ -41,7 +40,6 @@ class TransactionPacketHandler final : public PacketHandler {
   std::shared_ptr<TransactionManager> trx_mgr_;
 
   // FOR TESTING ONLY
-  std::shared_ptr<DagBlockManager> dag_blk_mgr_;
   std::shared_ptr<TestState> test_state_;
 
   std::atomic<uint64_t> received_trx_count_{0};
