+++ tests/util_test/conf/conf_taraxa3.json
@@ -22,7 +22,6 @@
   },
   "max_transactions_pool_warn" : 0,
   "max_transactions_pool_drop" : 0,
-  "max_block_queue_warn" : 0,
   "logging": {
     "configurations": [
       {
